+++ lib/openhab/core/actions.rb
@@ -7,21 +7,19 @@ module OpenHAB
     #
     # All openHAB's actions including those provided by add-ons are available, notably:
     # * Audio
-    # * Voice
-    # * Things
     # * Ephemeris
     # * Exec
     # * HTTP
     # * Ping
+    # * Semantics (see {Items::Semantics})
+    # * Voice
     #
     # From add-ons, e.g.:
-    # * Transformation
+    # * NotificationAction (from
+    #   [openHAB Cloud Connector](https://www.openhab.org/addons/integrations/openhabcloud/);
+    #   see {notify notify})
     # * PersistenceExtensions (see {Items::Persistence})
-    # * NotificationAction (from [openHAB Cloud Connector](https://www.openhab.org/addons/integrations/openhabcloud/))
-    #
-    # Global actions are available as "global" methods on {OpenHAB::DSL}, or
-    # explicitly from this {Actions} module, or you can explicitly reference a
-    # specific action.
+    # * Transformation
     #
     # Thing-specific actions can be accessed from the {Things::Thing Thing} object.
     # See {Things::Thing#actions Thing#actions}.
@@ -29,15 +27,15 @@ module OpenHAB
     # @example Run the TTS engine and output to the default audio sink. For more information see [Voice](https://www.openhab.org/docs/configuration/multimedia.html#voice)
     #   rule 'Say the time every hour' do
     #     every :hour
-    #     run { say "The time is #{TimeOfDay.now}" }
+    #     run { Voice.say "The time is #{TimeOfDay.now}" }
     #   end
     #
     #   rule 'Play an audio file' do
     #     every :hour
-    #     run { play_sound "beep.mp3", volume: 100 }
+    #     run { Audio.play_sound "beep.mp3", volume: 100 }
     #   end
     #
-    #   play_stream 'example.com'
+    #   Audio.play_stream 'example.com'
     #
     # @example Send a broadcast notification via the openHAB Cloud
     #   rule 'Send an alert' do
@@ -49,20 +47,12 @@ module OpenHAB
     #   rule 'Run a command' do
     #     every :day
     #     run do
-    #       execute_command_line('/bin/true')
-    #     end
-    #   end
-    #
-    # @example Execute an external command, referencing the Exec module directly
-    #   rule 'Run a command' do
-    #     every :day
-    #     run do
-    #       OpenHAB::Core::Actions::Exec::execute_command_line('/bin/true')
+    #       Exec.execute_command_line('/bin/true')
     #     end
     #   end
     #
     # @example Run a transformation
-    #   transform("MAP", "myfan.map", "0")
+    #   Transformation.transform("MAP", "myfan.map", "0")
     #
     module Actions
       OSGi.services("org.openhab.core.model.script.engine.action.ActionService")&.each do |service|
@@ -110,69 +100,6 @@ module OpenHAB
         else
           NotificationAction.send_broadcast_notification(msg.to_s, icon&.to_s, severity&.to_s)
         end
-      end
-
-      #
-      # Say text via openHAB Text-To-Speech service, Voice.say()
-      #
-      # @param text [String] The text to say
-      # @param voice [String] Specify a particular voice to use
-      # @param sink [String] Specify a particular sink to output the speech
-      # @param volume [PercentType] Specify the volume for the speech
-      #
-      # @return [void]
-      #
-      def say(text, voice: nil, sink: nil, volume: nil)
-        volume = PercentType.new(volume) unless volume.is_a?(PercentType) || volume.nil?
-        Voice.say(text.to_s, voice&.to_s, sink&.to_s, volume)
-      end
-
-      #
-      # Play an audio file via openHAB sound service, Audio.playSound()
-      #
-      # @param filename [String] The sound file to play
-      # @param sink [String] Specify a particular sink to output the speech
-      # @param volume [PercentType] Specify the volume for the speech
-      #
-      # @return [void]
-      #
-      def play_sound(filename, sink: nil, volume: nil)
-        volume = PercentType.new(volume) unless volume.is_a?(PercentType) || volume.nil?
-        Audio.play_sound(sink&.to_s, filename.to_s, volume)
-      end
-
-      #
-      # Play an audio stream from an URL to the given sink(s). Set url to nil if streaming should be stopped
-      #
-      # @param [String] url The URL of the audio stream
-      # @param [String] sink The audio sink, or nil to use the default audio sink
-      #
-      # @return [void]
-      #
-      def play_stream(url, sink: nil)
-        Audio.play_stream(sink&.to_s, url)
-      end
-
-      #
-      # Delegate missing methods to any available global actions.
-      #
-      def method_missing(method, *args, &block)
-        Actions.constants.each do |constant|
-          mod = Actions.const_get(constant)
-          return mod.public_send(method, *args, &block) if mod.respond_to?(method)
-        end
-
-        super
-      end
-
-      # @!visibility private
-      def respond_to_missing?(method_name, _include_private = false)
-        Actions.constants.each do |constant|
-          mod = Actions.const_get(constant)
-          return true if mod.respond_to?(method_name)
-        end
-
-        super
       end
     end
   end