+++ lib/openhab/core/actions/audio.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+module OpenHAB
+  module Core
+    module Actions
+      # @see https://www.openhab.org/docs/configuration/multimedia.html#actions-2 Audio Actions
+      class Audio
+        class << self
+          #
+          # Play an audio file via openHAB sound service, Audio.playSound()
+          #
+          # @param filename [String] The sound file to play
+          # @param sink [String] Specify a particular sink to output the speech
+          # @param volume [PercentType] Specify the volume for the speech
+          #
+          # @return [void]
+          #
+          def play_sound(filename, sink: nil, volume: nil)
+            volume = PercentType.new(volume) unless volume.is_a?(PercentType) || volume.nil?
+            playSound(sink&.to_s, filename.to_s, volume)
+          end
+
+          #
+          # Play an audio stream from an URL to the given sink(s). Set url to nil if streaming should be stopped
+          #
+          # @param [String] url The URL of the audio stream
+          # @param [String] sink The audio sink, or nil to use the default audio sink
+          #
+          # @return [void]
+          #
+          def play_stream(url, sink: nil)
+            playStream(sink&.to_s, url)
+          end
+        end
+      end
+    end
+  end
+end